Handle index overrun in linkage lists, remove unused imports, and add docs

+++ python/lsst/solsys/pipe/loadBalanceLinkageBundles.py
@@ -20,57 +20,50 @@
 # see <https://www.lsstcorp.org/LegalNotices/>.
 #
 
-__all__ = ("LoadBalanceConfig",
-           "LoadBalanceTask")
+__all__ = ("LoadBalanceConfig", "LoadBalanceTask")
+
+import logging
 
 import astropy.table
-import astropy.units as u
-from collections import defaultdict
-from datetime import datetime
-import heliolinx.heliolinx as hl
-import logging
-import numpy as np
-import pandas as pd
-
-import lsst.pipe.base
-from lsst.pex.config import Config, ConfigField, ConfigurableField, Field
+from lsst.pex.config import Field
 from lsst.pipe.base import (
-    connectionTypes,
     NoWorkFound,
     PipelineTask,
     PipelineTaskConfig,
     PipelineTaskConnections,
-    Struct
+    Struct,
+    connectionTypes,
 )
-
 
 _LOG = logging.getLogger(__name__)
 
-class LoadBalanceConnections(PipelineTaskConnections,
-                             dimensions=["instrument", "day_obs", "ssp_hypothesis_table"]):
+
+class LoadBalanceConnections(
+    PipelineTaskConnections, dimensions=["instrument", "day_obs", "ssp_hypothesis_table"]
+):
     sspLinkageList = connectionTypes.Input(
-        doc="",
+        doc="List of linkage tables to be load-balanced.",
         dimensions=["day_obs", "ssp_hypothesis_table", "ssp_hypothesis_bundle"],
         storageClass="ArrowAstropy",
         name="ssp_linkages",
         multiple=True,
     )
     sspLinkageSourceList = connectionTypes.Input(
-        doc="",
+        doc="List of linkage source tables corresponding to the linkage tables.",
         dimensions=["day_obs", "ssp_hypothesis_table", "ssp_hypothesis_bundle"],
         storageClass="ArrowAstropy",
         name="ssp_linkage_sources",
         multiple=True,
     )
     sspLoadBalancedLinkages = connectionTypes.Output(
-        doc="",
+        doc="List of load-balanced linkage tables.",
         dimensions=["day_obs", "ssp_hypothesis_table", "ssp_balanced_index"],
         storageClass="ArrowAstropy",
         name="ssp_balanced_linkages",
         multiple=True,
     )
     sspLoadBalancedLinkageSources = connectionTypes.Output(
-        doc="",
+        doc="List of load-balanced linkage source tables.",
         dimensions=["day_obs", "ssp_hypothesis_table", "ssp_balanced_index"],
         storageClass="ArrowAstropy",
         name="ssp_balanced_linkage_sources",
@@ -79,11 +72,8 @@ class LoadBalanceConnections(PipelineTaskConnections,
 
 
 class LoadBalanceConfig(PipelineTaskConfig, pipelineConnections=LoadBalanceConnections):
-    num_linkrefine_indices = Field(
-        dtype=int,
-        default=10,
-        doc="Number of linkRefine quanta / output datasets"
-    )
+    num_linkrefine_indices = Field(dtype=int, default=10, doc="Number of linkRefine quanta / output datasets")
+
 
 class LoadBalanceTask(PipelineTask):
     ConfigClass = LoadBalanceConfig
@@ -100,35 +90,56 @@ class LoadBalanceTask(PipelineTask):
             butlerQC.put(outputs.sspLoadBalancedLinkageSourceList[i], dataId)
 
     def run(self, sspLinkageList, sspLinkageSourceList):
-        """doc string
-           here
+        """Load balance the linkage tables across the specified number of
+        indices.
+
+        Parameters
+        ----------
+        sspLinkageList : `list` of `astropy.table.Table`
+            List of linkage tables to be load-balanced.
+        sspLinkageSourceList : `list` of `astropy.table.Table`
+            List of linkage source tables corresponding to the linkage tables.
+
+        Returns
+        -------
+        `Struct`
+            A structure containing two lists:
+            - ``sspLoadBalancedLinkageList``: List of load-balanced linkage
+              tables.
+            - ``sspLoadBalancedLinkageSourceList``: List of load-balanced
+              linkage source tables.
         """
         if len(sspLinkageList) == 0:
             raise NoWorkFound
 
         n_ind = self.config.num_linkrefine_indices
-        _LOG.info(f'Concatenating {len(sspLinkageList)} linkage tables')
+        _LOG.info(f"Concatenating {len(sspLinkageList)} linkage tables")
         n = 0
         for i in range(len(sspLinkageList)):
             sspLinkage = sspLinkageList[i]
-            sspLinkage['clusternum'] += n
+            sspLinkage["clusternum"] += n
             sspLinkageList[i] = sspLinkage
             sspLinkageSource = sspLinkageSourceList[i]
-            sspLinkageSource['i1'] += n
+            sspLinkageSource["i1"] += n
             sspLinkageSourceList[i] = sspLinkageSource
             n += len(sspLinkage)
         sspLinkage = astropy.table.vstack(sspLinkageList)
         sspLinkageSource = astropy.table.vstack(sspLinkageSourceList)
-        sspLinkage['loadBalanceIndex'] = sspLinkage['clusternum'] % n_ind
-        sspLinkageSource['loadBalanceIndex'] = sspLinkageSource['i1'] % n_ind
-        sspLoadBalancedLinkageList = [t for t in sspLinkage.group_by('loadBalanceIndex').groups]
-        sspLoadBalancedLinkageSourceList = [t for t in sspLinkageSource.group_by('loadBalanceIndex').groups]
+        sspLinkage["loadBalanceIndex"] = sspLinkage["clusternum"] % n_ind
+        sspLinkageSource["loadBalanceIndex"] = sspLinkageSource["i1"] % n_ind
+        sspLoadBalancedLinkageList = [t for t in sspLinkage.group_by("loadBalanceIndex").groups]
+        sspLoadBalancedLinkageSourceList = [t for t in sspLinkageSource.group_by("loadBalanceIndex").groups]
         for i in range(n_ind):
-            sspLoadBalancedLinkageList[i]['clusternum'] //= n_ind
-            sspLoadBalancedLinkageSourceList[i]['i1'] //= n_ind
-            sspLoadBalancedLinkageList[i].remove_column('loadBalanceIndex')
-            sspLoadBalancedLinkageSourceList[i].remove_column('loadBalanceIndex')
+            if len(sspLoadBalancedLinkageList) <= i:
+                sspLoadBalancedLinkageList.append(astropy.table.Table())
+                sspLoadBalancedLinkageSourceList.append(astropy.table.Table())
+            else:
+                sspLoadBalancedLinkageList[i]["clusternum"] //= n_ind
+                sspLoadBalancedLinkageSourceList[i]["i1"] //= n_ind
+                sspLoadBalancedLinkageList[i].remove_column("loadBalanceIndex")
+                sspLoadBalancedLinkageSourceList[i].remove_column("loadBalanceIndex")
 
-        return Struct(sspLoadBalancedLinkageList = sspLoadBalancedLinkageList,
-                      sspLoadBalancedLinkageSourceList = sspLoadBalancedLinkageSourceList
-                     )
+        return Struct(
+            sspLoadBalancedLinkageList=sspLoadBalancedLinkageList,
+            sspLoadBalancedLinkageSourceList=sspLoadBalancedLinkageSourceList,
+        )
